Update tests and fix patches/mocks
